+++ CHANGELOG.md
@@ -342,6 +342,7 @@ in the naming of release branches.
 - Fixed typo in makeHashWithExplicitProxys phantom type (indexl to index). #3072
 - Fixed the incorrect conversion of the validity interval's upper bound in `transVITime` (fixes #3043). #3200
 - Enforce that the CostModel deserializers expect a specific length prior to version 9.
+- Starting in version 9, duplicate keys in CBOR maps are not longer allowed
 
 ## Release branch 1.1.x